requirements.txt
flask
requests
python-dotenv
google-generativeai
matplotlib
Pillow
seaborn
reportlab
opencv-python
gunicorn
